포스팅 목록 불러와서 컨텐츠 부분 구현 완료. 댓글 해야함.

+++ taek-story-spring-cloud/file-service/src/main/java/com/example/fileservice/mapper/FileMapper.java
@@ -10,7 +10,7 @@ public interface FileMapper {
 	int updateFileMaster(FileMaster fileMaster);
 
 	int insertFileDetail(FileDetail fileDetail);
-	int deleteFileDetailByMasterSeq(Integer fileMasterSeq);
+	int deleteFileDetailByMasterSeq(Long fileMasterSeq);
 
-	FileDetail selectFileDetailOneByMasterSeq(Integer seq);
+	FileDetail selectFileDetailOneByMasterSeq(Long seq);
 }
