+++ taek-story-spring-cloud/posting-service/src/main/java/com/example/postingservice/feign/FileServiceClient.java
@@ -17,10 +17,10 @@ public interface FileServiceClient {
 	@PostMapping(value = "/file/upload/posting", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
 	FileMaster uploadFilePosting(
 			@RequestPart("file") MultipartFile file,
-			@RequestParam(value = "fileMasterSeq", required = false) Integer fileMasterSeq,
+			@RequestParam(value = "fileMasterSeq", required = false) Long fileMasterSeq,
 			@RequestParam("fileType") String fileType,
 			@RequestParam("ip") String ip,
-			@RequestParam("userSeq") Integer userSeq
+			@RequestParam("userSeq") Long userSeq
 	);
 
 }
